Move QuarkusUnitTest function to new class but maintain old function
